+++ Assets/uni-csv/Runtime/CsvColumnAttribute.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace UniCSV
+{
+    [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field)]
+    public sealed class CsvColumnAttribute : Attribute
+    {
+        public string Name { get; }
+
+        public CsvColumnAttribute(string name)
+        {
+            Name = name;
+        }
+    }
+}