+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2024-10-06
+
+### Breaking changes
+
+- Renamed the base namespace to `UniCSV`.
+
+- Renamed the primary static class to `CsvParser`.
+
+## Changed
+
+- Moved some parts of the code to their own utility classes and files.
+
 ## [1.1.0] - 2024-10-06
 
 ### Added